Fix build

+++ belle-sip/src/channel.c
@@ -419,7 +419,7 @@ static int acquire_body_simple(belle_sip_channel_t *obj, int end_of_stream){
 	if (end_of_stream ||  belle_sip_body_handler_get_transfered_size(bh)>=content_length){
 		/*great body completed*/
 		belle_sip_message("channel [%p] read [%i] bytes of body from [%s:%i]"	,obj
-			,content_length
+			,(int)content_length
 			,obj->peer_name
 			,obj->peer_port);
 		belle_sip_channel_message_ready(obj);
